rework 'worker.js' to expect the full data, not just the filename
